Moved code out of head of ares_help_info.
